requirements.txt
fastapi
uvicorn
pillow
torch
openai-clip